# backend/api/api.yml
openapi: "3.0.0"
info:
  version: "1.20"
  title: Multicloud API
#produces:
#  - "application/json"
#basePath: "/api"
servers:
  - url: /api

# Paths supported by the server application
paths:
  /aws/vm:
    post:
      operationId: "aws.vm.createVM"
      tags:
        - "AWS-vm"
      summary: "Create an EC2 (VM)"
      description: "Create an EC2 (VM)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMRequest'
      responses:
        200:
          description: "Successful created VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    put:
      operationId: "aws.vm.loadAllVM"
      tags:
        - "AWS-vm"
      summary: "Loads All VMs"
      description: "Loads all VMS"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful loaded all VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMResponseFullist'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    delete:
      operationId: "aws.vm.deleteVM"
      tags:
        - "AWS-vm"
      summary: "Deletes an EC2 (VM)"
      description: "Deletes an EC2 (VM)"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMRequestShort'
      responses:
        200:
          description: "Successful deleted VM"
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
  /aws/vm/{instanceId}:
    post:
      operationId: "aws.vm.getVM"
      tags:
        - "AWS-vm"
      summary: "Get the VM based on provided instanceId"
      description : "Get the VM based on provided instanceId"
      parameters:
        - $ref: '#/components/parameters/instanceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful read VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    patch:
      operationId: "aws.vm.stopVM"
      tags:
        - "AWS-vm"
      summary: "stop the VM based on provided instanceId"
      description: "stop the VM based on provided instanceId"
      parameters:
        - $ref: '#/components/parameters/instanceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful stopped VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    put:
      operationId: "aws.vm.startVM"
      tags:
        - "AWS-vm"
      summary: "Starts the VM based on provided instanceId"
      description: "Starts the VM based on provided instanceId"
      parameters:
        - $ref: '#/components/parameters/instanceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful started VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/State'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
...
  /azure/resource:
    post:
      operationId: "azure.resourcegroup.createOrUpdateResourceGroup"
      tags:
        - "AZURE-resourcegroup"
      summary: "Creates or Updates a ResourceGroup"
      description: "Creates or Updates a ResourceGroup"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceGroupRequest'
      responses:
        200:
          description: "Successful created ResourceGroup"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGroupResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    put:
      operationId: "azure.resourcegroup.getAllResourceGroups"
      tags:
        - "AZURE-resourcegroup"
      summary: "Gets all ResourceGroups"
      description: "Gets all ResourceGroups"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      responses:
        200:
          description: "Successful read ResourceGroups"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGroupListResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    delete:
      operationId: "azure.resourcegroup.deleteResourceGroup"
      tags:
        - "AZURE-resourcegroup"
      summary: "deletes a resourcegroup from azure"
      description: "deletes a resourcegroup from azure"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourceGroupRequest'
      responses:
        200:
          description: "Successful deleted ResourceGroup"
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
  /azure/resource/{groupName}:
    post:
      operationId: "azure.resourcegroup.getAResourceGroup"
      tags:
        - "AZURE-resourcegroup"
      summary: "Get a resourcegroup from azure"
      description: "Get a resourcegroup from azure"
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRequest'
      parameters:
        - $ref: '#/components/parameters/groupName'
      responses:
        200:
          description: "Successful read ResourceGroup"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourceGroupResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
  /azure/vm:
    post:
      operationId: "azure.vm.createaVM"
      tags:
        - "AZURE-vm"
      summary: "Creates a VM"
      description: "Creates a VM"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AzureVMRequest'
      responses:
        200:
          description: "Successful created VM"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AzureVMResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    patch:
      operationId: "azure.vm.getAllVMS"
      tags:
        - "AZURE-vm"
      summary: "gets all VMs"
      description: "gets all VMs"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NetworkListRequest'
      responses:
        200:
          description: "Successful read VMs"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VMListResponse'
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    delete:
      operationId: "azure.vm.deleteVM"
      tags:
        - "AZURE-vm"
      summary: "deletes a vm from azure"
      description: "deletes a vm from azure"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMSmallRequest'
      responses:
        200:
          description: "Successful deleted VM"
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
    put:
      operationId: "azure.vm.startVM"
      tags:
        - "AZURE-vm"
      summary: "starts a VM"
      description: "starts a VM"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VMSmallRequest'
      responses:
        200:
          description: "Successful started VM"
        403:
          description: "Not authorized"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NotAuthorizedResponse'
...
components:
  parameters:
...
    groupName:
      name: groupName
      in: path
      required: true
      description: groupName
      schema:
        type: string
...
    instanceId:
      name: instanceId
      in: path
      required: true
      description: InstanceId
      schema:
        type: string
...
  schemas:
    TokenRequest:
      type: object
      properties:
        token:
          type: string
    NotAuthorizedResponse:
      type: object
      properties:
        detail:
          type: string
        status:
          type: number
        title:
          type: string
        type:
          type: string
...
    VMSmallRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
        vmname:
          type: string
    VMListResponse:
      type: array
      items:
        $ref: '#/components/schemas/AzureVMResponse'
    AzureVMResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        provisioning_state:
          type: string
        instance_view:
          type: object
          properties:
            platform_update_domain:
              type: number
            platform_fault_domain:
              type: number
            computer_name:
              type: string
            os_name:
              type: string
            os_version:
              type: string
            hyper_vgeneration:
              type: string
            rdp_thumb_print:
              type: string
            vm_agent:
              type: object
              properties:
                vm_agent_version:
                  type: string
                extension_handlers:
                  type: array
                  items:
                    type: object
                    properties:
                      type:
                        type: string
                      type_handler_version:
                        type: string
                      status:
                        $ref: '#/components/schemas/Statuses'
                statuses:
                  type: array
                  items:
                    $ref: '#/components/schemas/Statuses'
            maintenance_redeploy_status:
              type: object
              properties:
                is_customer_initiated_maintenance_allowed:
                  type: boolean
                pre_maintenance_window_start_time:
                  type: string
                  format: date-time
                pre_maintenance_window_end_time:
                  type: string
                  format: date-time
                maintenance_window_start_time:
                  type: string
                  format: date-time
                maintenance_window_end_time:
                  type: string
                  format: date-time
                last_operation_result_code:
                  type: string
                last_operation_message:
                  type: string
            disks:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  encryption_settings:
                    type: array
                    items:
                      type: object
                      properties:
                        disk_encryption_key:
                          type: object
                          properties:
                            secret_url:
                              type: string
                            source_vault:
                              type: object
                              properties:
                                id:
                                  type: string
                        key_encryption_key:
                          type: object
                          properties:
                            key_url:
                              type: string
                            source_vault:
                              type: object
                              properties:
                                id:
                                  type: string
                        enabled:
                          type: boolean
                  statuses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Statuses'
            extensions:
              type: array
              items:
                type: object
                properties:
                  name:
                    type: string
                  type:
                    type: string
                  type_handler_version:
                    type: string
                  substatuses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Statuses'
                  statuses:
                    type: array
                    items:
                      $ref: '#/components/schemas/Statuses'
            boot_diagnostics:
              type: object
              properties:
                console_screenshot_blob_uri:
                  type: string
                serial_console_log_blob_uri:
                  type: string
                status:
                  $ref: '#/components/schemas/Statuses'
            statuses:
              type: array
              items:
                $ref: '#/components/schemas/Statuses'
        vm_id:
          type: string
    Statuses:
      type: object
      properties:
        code:
          type: string
        level:
          type: string
        display_status:
          type: string
        message:
          type: string
        time:
          type: string
          format: date-time
    AzureVMRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
        vmname:
          type: string
        username:
          type: string
        password:
          type: string
        vmsize:
          type: string
        publishername:
          type: string
        offername:
          type: string
        skuname:
          type: string
        version:
          type: string
        nicid:
          type: string
    ResourceGroupRequest:
      type: object
      properties:
        token:
          type: string
        groupName:
          type: string
    ResourceGroupListResponse:
      type: array
      items:
        $ref: '#/components/schemas/ResourceGroupResponse'
    ResourceGroupResponse:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
...
    NetworkListRequest:
      type: object
      properties:
        token:
          type: string
        resourcegroupname:
          type: string
...
    VMRequest:
      type: object
      properties:
        token:
          type: string
        imageId:
          type: string
        instanceType:
          type: string
        keyName:
          type: string
        securityGroups:
          type: array
          items:
            type: string
        vmname:
          type: string
    VMRequestShort:
      type: object
      properties:
        token:
          type: string
        instanceId:
          type: string
    VMResponse:
      type: object
      properties:
        instanceId:
          type: string
    State:
      type: object
      properties:
        State:
          type: string
...
    VMResponseFullist:
      type: object
      properties:
        instances:
          type: array
          items:
            $ref: '#/components/schemas/VMResponseFull'
    VMResponseFull:
      type: object
      properties:
        Groups:
          type: array
          items:
            type: object
            properties:
              GroupName:
                type: string
              GroupId:
                type: string
        Instances:
          type: array
          items:
            type: object
            properties:
              AmiLaunchIndex:
                type: integer
              ImageId:
                type: string
              InstanceId:
                type: string
              InstanceType:
                type: string
              KernelId:
                type: string
              KeyName:
                type: string
              LaunchTime:
                type: string
              Monitoring:
                type: object
                properties:
                  State:
                    type: string
              Placement:
                type: object
                properties:
                  AvailabilityZone:
                    type: string
                  Affinity:
                    type: string
                  GroupName:
                    type: string
                  PartitionNumber:
                    type: integer
                  HostId:
                    type: string
                  Tenancy:
                    type: string
                  SpreadDomain:
                    type: string
                  HostResourceGroupArn:
                    type: string
              Platform:
                type: string
              PrivateDnsName:
                type: string
              PrivateIpAddress:
                type: string
              ProductCodes:
                type: array
                items:
                  type: object
                  properties:
                    ProductCodeId:
                      type: string
                    ProductCodeType:
                      type: string
              PublicDnsName:
                type: string
              PublicIpAddress:
                type: string
              RamdiskId:
                type: string
              State:
                type: object
                properties:
                  Code:
                    type: integer
                  Name:
                    type: string
              StateTransitionReason:
                type: string
              SubnetId:
                type: string
              VpcId:
                type: string
              Architecture:
                type: string
              BlockDeviceMappings:
                type: array
                items:
                  type: object
                  properties:
                    DeviceName:
                      type: string
                    Ebs:
                      type: object
                      properties:
                        AttachTime:
                          type: string
                        DeleteOnTermination:
                          type: boolean
                        Status:
                          type: string
                        VolumeId:
                          type: string
              ClientToken:
                type: string
              EbsOptimized:
                type: boolean
              EnaSupport:
                type: boolean
              Hypervisor:
                type: string
              IamInstanceProfile:
                type: object
                properties:
                  Arn:
                    type: string
                  Id:
                    type: string
              InstanceLifecycle:
                type: string
              ElasticGpuAssociations:
                type: array
                items:
                    type: object
                    properties:
                      ElasticGpuId:
                        type: string
                      ElasticGpuAssociationId:
                        type: string
                      ElasticGpuAssociationState:
                        type: string
                      ElasticGpuAssociationTime:
                        type: string
              ElasticInferenceAcceleratorAssociations:
                type: array
                items:
                  type: object
                  properties:
                    ElasticInferenceAcceleratorArn:
                      type: string
                    ElasticInferenceAcceleratorAssociationId:
                      type: string
                    ElasticInferenceAcceleratorAssociationState:
                      type: string
                    ElasticInferenceAcceleratorAssociationTime:
                      type: string
              NetworkInterfaces:
                type: array
                items:
                  type: object
                  properties:
                    Association:
                      type: object
                      properties:
                        IpOwnerId:
                          type: string
                        PublicDnsName:
                          type: string
                        PublicIp:
                          type: string
                    Attachment:
                      type: object
                      properties:
                        AttachTime:
                          type: string
                        AttachmentId:
                          type: string
                        DeleteOnTermination:
                          type: boolean
                        DeviceIndex:
                          type: integer
                        Status:
                          type: string
                    Description:
                      type: string
                    Groups:
                      type: array
                      items:
                        type: object
                        properties:
                          GroupName:
                            type: string
                          GroupId:
                            type: string
                    Ipv6Addresses:
                      type: array
                      items:
                        type: object
                        properties:
                          Ipv6Address:
                            type: string
                    MacAddress:
                      type: string
                    NetworkInterfaceId:
                      type: string
                    OwnerId:
                      type: string
                    PrivateDnsName:
                      type: string
                    PrivateIpAddress:
                      type: string
                    PrivateIpAddresses:
                      type: array
                      items:
                        type: object
                        properties:
                          Association:
                            type: object
                            properties:
                              IpOwnerId:
                                type: string
                              PublicDnsName:
                                type: string
                              PublicIp:
                                type: string
                          Primary:
                            type: boolean
                          PrivateDnsName:
                            type: string
                          PrivateIpAddress:
                            type: string
                    SourceDestCheck:
                      type: boolean
                    Status:
                      type: string
                    SubnetId:
                      type: string
                    VpcId:
                      type: string
                    InterfaceType:
                      type: string
              OutpostArn:
                type: string
              RootDeviceName:
                type: string
              RootDeviceType:
                type: string
              SecurityGroups:
                type: array
                items:
                  type: object
                  properties:
                    GroupName:
                      type: string
                    GroupId:
                      type: string
              SourceDestCheck:
                type: boolean
              SpotInstanceRequestId:
                type: string
              SriovNetSupport:
                type: string
              StateReason:
                type: object
                properties:
                  Code:
                    type: string
                  Message:
                    type: string
              Tags:
                type: array
                items:
                  type: object
                  properties:
                    Key:
                      type: string
                    Value:
                      type: string
              VirtualizationType:
                type: string
              CpuOptions:
                type: object
                properties:
                  CoreCount:
                    type: integer
                  ThreadsPerCore:
                    type: integer
              CapacityReservationId:
                type: string
              CapacityReservationSpecification:
                type: object
                properties:
                  CapacityReservationPreference:
                    type: string
                  CapacityReservationTarget:
                    type: object
                    properties:
                      CapacityReservationId:
                        type: string
              HibernationOptions:
                type: object
                properties:
                  Configured:
                    type: boolean
              Licenses:
                type: array
                items:
                  type: object
                  properties:
                    LicenseConfigurationArn:
                      type: string
              MetadataOptions:
                type: object
                properties:
                  State:
                    type: string
                  HttpTokens:
                    type: string
                  HttpPutResponseHopLimit:
                    type: integer
                  HttpEndpoint:
                    type: string
        OwnerId:
          type: string
        RequesterId:
          type: string
        ReservationId:
          type: string
...